Additional test case
